+++ CHANGELOG.md
@@ -7,8 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Configurable `SendRate` for deterministic replication. Use `SendRate::Once` to send only the initial value, or `SendRate::Periodic` to only sync the state periodically.
+- `AppRuleExt::replicate_with_priority` to configure replication rule priority.
+
 ### Changed
 
+- `AppRuleExt::replicate_with` now accepts `IntoReplicationRule` trait that allows to define rules with multiple components.
+- Rename `GroupReplication` into `BundleReplication`.
+- Rename `AppRuleExt::replicate_group` into `AppRuleExt::replicate_bundle`.
+- `ReplicationRule` now stores `Vec<ComponentRule>` instead of `Vec<(ComponentId, FnsId)>`
+- `RuleFns` now available from prelude.
 - Rules created with the same priority now evaluated in their creation order.
 
 ### Removed